Backend/requirements.txt
fastapi
groq
PyJWT
Pillow
pandas
pytesseract
python-dotenv
python-multipart
SQLAlchemy
uvicorn
Werkzeug
